requirements.txt
asgiref==3.4.1; python_version >= "3.6"
boto3==1.20.23; python_version >= "3.6"
botocore==1.23.23; python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; python_version >= "3.6"
charset-normalizer==2.0.9; python_full_version >= "3.6.0" and python_version >= "3"
cryptography==36.0.0; python_version >= "3.6"
defusedxml==0.7.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
django-allauth==0.43.0
django-cors-headers==3.10.1; python_version >= "3.6"
django-crispy-forms==1.10.0; python_version >= "3.5"
django-debug-toolbar==3.2; python_version >= "3.6"
django-environ==0.8.1; python_version >= "3.4" and python_version < "4"
django-storages==1.12.3; python_version >= "3.5"
django==3.1.4; python_version >= "3.6"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
gunicorn==20.1.0; python_version >= "3.5"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
psycopg2-binary==2.9.2; python_version >= "3.6"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pyjwt==2.3.0; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python3-openid==3.2.0
pytz==2021.3; python_version >= "3.6"
requests-oauthlib==1.3.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests==2.26.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
s3transfer==0.5.0; python_version >= "3.6"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
whitenoise==5.2.0; python_version >= "3.5" and python_version < "4"
